Rename(Modal): 모달 컴포넌트 폴더 생성 및 이동
